Move fbUtils to src/utilities and rename
