+++ package.json
@@ -56,7 +56,7 @@
     "uuid": "^3.3.2"
   },
   "devDependencies": {
-    "browser-sync": "^2.26.3",
+    "browser-sync": "^2.26.5",
     "cross-env": "^5.2.0",
     "eslint": "^5.11.0",
     "eslint-config-airbnb-base": "^13.1.0",
